flask-server/requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
Pillow>=9.0.0
numpy>=1.24.0
gunicorn>=20.1.0
opencv-python-headless>=4.7.0.72
itsdangerous>=2.0.0
tqdm>=4.64.0 
flask_cors>=3.0.10
cloudinary>=1.36.0
requests>=2.31.0
torch>=2.0.0
torchvision>=0.15.0
scikit-image>=0.19.3
scipy>=1.9.0
ultralytics>=8.1.0
transformers>=4.30.0